// src/enums/page-view-type.enum.ts
export enum PageViewType {
  main = 'main',
  invests = 'invests',
  article = 'article',
  articles = 'articles',
  doubleArticle = 'doubleArticle',
  aboutArticleMapsView = 'aboutArticleMapsView',
  rodoArticle = 'rodoArticle',
  contact = 'contact',
  department = 'department',
  notices = 'notices',
  blank = 'blank',
}